+++ app/Models/Booking.php
@@ -181,12 +181,12 @@ class Booking extends Model
 
     public function airportDep()
     {
-        return $this->hasOne(Airport::class, 'icao', 'dep');
+        return $this->hasOne(Airport::class, 'id', 'dep');
     }
 
     public function airportArr()
     {
-        return $this->hasOne(Airport::class, 'icao', 'arr');
+        return $this->hasOne(Airport::class, 'id', 'arr');
     }
 
     public function event()