+++ database/factories/AirportLinkFactory.php
@@ -4,8 +4,8 @@ use Faker\Generator as Faker;
 
 $factory->define(App\Models\AirportLink::class, function (Faker $faker) {
     return [
-        'icao_airport' => function () {
-            return factory(App\Models\Airport::class)->create()->icao;
+        'airport_id' => function () {
+            return factory(App\Models\Airport::class)->create()->id;
         },
         'airportLinkType_id' => 4,
         'url' => $faker->url,